+++ src/Param/InputInterface.php
@@ -13,4 +13,20 @@ interface InputInterface extends ParamHolderInterface
      * @return InputInterface
      */
     public static function fromOutput(OutputInterface $input): InputInterface;
+
+    /**
+     * @return string
+     */
+    public function getEvent(): string;
+
+    /**
+     * @return boolean
+     */
+    public function hasEvent(): bool;
+
+    /**
+     * @param  string $event
+     * @return InputInterface
+     */
+    public function withEvent(string $event): InputInterface;
 }